Restrict availability of helper method to just where it's needed
This makes it easier to find, and helps avoid it being called
in places it's not meant to.
